State bug
After investigating some valuable drops not sending webhook reports from some users, I am currently convinced it has to do with some state issues. I have moved the state logic to when sending webhooks / logs. I have also added a suite of unit tests which should have already been there to begin with.

+++ src/main/java/com/fauxbingo/handlers/ValuableDropHandler.java
@@ -68,7 +68,7 @@ public class ValuableDropHandler implements EventHandler<ChatMessage>
 		
 		if (matcher.matches())
 		{
-			int valuableDropValue = Integer.parseInt(matcher.group(2).replaceAll(",", ""));
+			long valuableDropValue = Long.parseLong(matcher.group(2).replaceAll(",", ""));
 			
 			if (valuableDropValue >= config.valuableDropThreshold())
 			{
